Fix file delete error in test. Related to Issue [WTI-242]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@1846 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: a433aab3c170fb6b100720612a3562619387539a

+++ Core/Components/test/Core.Components.DotSpatial.Test/Data/MapDataTest.cs
@@ -19,7 +19,10 @@ namespace Core.Components.DotSpatial.Test.Data
 
             RenameFile(newPath, oldPath);
 
-            File.Delete(oldPath);
+            if (File.Exists(oldPath))
+            {
+                File.Delete(oldPath);
+            }
         }
 
         [Test]
